Use an int64_t instead of size_t for nCoinCacheUsage
Because nCoinCacheUsage is size_t, then on 32bit systems when
we assign it a value from 64bit int then we end up with an
incorrect value. This happens during init when the cache sizes
are being calculated.

+++ src/main.h
@@ -184,7 +184,7 @@ extern bool fRequireStandard;
 extern unsigned int nBytesPerSigOp;
 extern bool fCheckBlockIndex;
 extern bool fCheckpointsEnabled;
-extern size_t nCoinCacheUsage;
+extern int64_t nCoinCacheUsage;
 /** A fee rate smaller than this is considered zero fee (for relaying, mining and transaction creation) */
 extern CFeeRate minRelayTxFee;
 /** Absolute maximum transaction fee (in satoshis) used by wallet and mempool (rejects high fee in sendrawtransaction)
